split out dirtyIndicator Plugin
