+++ open-authoring/toolbar-plugins/DirtyIndicatorPlugin.tsx
@@ -1,0 +1,61 @@
+import styled, { css } from 'styled-components'
+import { FieldMeta } from 'tinacms'
+import { DesktopLabel } from '../../components/ui/inline/DesktopLabel'
+import { color } from '@tinacms/styles'
+
+export const DirtyIndicatorPlugin = formState => ({
+  __type: 'toolbar:status',
+  name: 'form-state-dirty',
+  props: {
+    dirty: formState.dirty,
+  },
+  component: FormStatus,
+})
+
+interface StatusLightProps {
+  warning?: boolean
+}
+
+const StatusLight = styled.span<StatusLightProps>`
+  display: inline-block;
+  width: 8px;
+  height: 8px;
+  border-radius: 8px;
+  margin-top: -1px;
+  background-color: #3cad3a;
+  border: 1px solid #249a21;
+  margin-right: 5px;
+  opacity: 0.5;
+
+  ${p =>
+    p.warning &&
+    css`
+      background-color: #e9d050;
+      border: 1px solid #d3ba38;
+      opacity: 1;
+    `};
+`
+
+const StatusMessage = styled.p`
+  font-size: 16px;
+  display: flex;
+  align-items: center;
+  color: ${color.grey(6)};
+  padding-right: 4px;
+`
+
+const FormStatus = ({ dirty }) => {
+  return (
+    <FieldMeta name={'Form Status'}>
+      {dirty ? (
+        <StatusMessage>
+          <StatusLight warning /> <DesktopLabel>Unsaved changes</DesktopLabel>
+        </StatusMessage>
+      ) : (
+        <StatusMessage>
+          <StatusLight /> <DesktopLabel>No changes</DesktopLabel>
+        </StatusMessage>
+      )}
+    </FieldMeta>
+  )
+}
